Made outline
Basic outline of website has been created, with homepage, and
animated scrolling for index.

I have not really tested the bar on mobile, and to be honest
I think Ill have to create an issue or soething with regards to that.

Images need to be replaced, and link to CV needs to be done.